requirements.txt
fastapi==0.111.2
uvicorn[standard]==0.23.2
pydantic==2.6.2
python-dotenv==1.0.1
rich==13.5.3
requests==2.32.0
streamlit==1.27.1

langchain==0.1.197
langchain-core==0.0.253
langchain-community==0.1.51
langchain-text-splitters==0.1.8
langchain-ollama==0.0.16
langsmith==0.0.31

chromadb==0.3.34
ollama==0.0.37
tiktoken==0.4.0
sentence-transformers==2.2.2
gitpython==3.1.32
aiofiles==23.1.0

# Optional for testing
pytest==7.4.0
pytest-asyncio==0.21.0